requirements.txt
Flask
Flask-Limiter
Flask-SocketIO
pycryptodome
pyjwt
werkzeug
flask_limiter.storage